Added edit address feature (Phase II)

+++ java-classes-and-objects/PracticeProgramming/src/main/java/com/mthree/academy/c458/vrishti/jco/exercises/address_book/controller/AddressBookController.java
@@ -38,6 +38,9 @@ public class AddressBookController {
                     findAllAddresses();
                     break;
                 case 6:
+                    editAddress();
+                    break;
+                case 7:
                     keepGoing = false;
                     break;
                 default:
@@ -69,11 +72,23 @@ public class AddressBookController {
 
         //Get id of address to remove
         long addressId = view.getAddressId();
-        //Remove the address
-        boolean success = dao.removeAddress(addressId);
-        //Feedback to user
-        if (success) view.printSuccessMessage("Removed address from address book.");
-        else view.printWarnMessage("Could not remove address from address book.");
+
+        //Get address
+        Address address = dao.getAddressById(addressId);
+
+        //View confirm removal
+        boolean confirmedDelete = view.confirmRemoveAddress(address);
+        if (confirmedDelete) {
+
+            //Remove the address
+            boolean success = dao.removeAddress(addressId);
+            //Feedback to user
+            if (success) view.printSuccessMessage("Removed address from address book.");
+            else view.printWarnMessage("Could not remove address from address book.");
+
+        } else {
+            view.printCancelledOperation();
+        }
     }
 
     private void findAddressesByLastName() {
@@ -103,6 +118,68 @@ public class AddressBookController {
         List<Address> addresses = dao.getAllAddresses();
         //Display all addresses
         view.displayAddresses("Address Book", addresses);
+    }
+
+    private void editAddress() {
+        view.printHeader("Edit Address");
+
+        //Address to update
+        long addressId = view.getAddressId();
+        Address address = dao.getAddressById(addressId);
+
+        //Run edit address submenu.
+        if (address != null)
+            runEditAddress(address);
+        else
+            view.printWarnMessage("Cannot update non-existent address.");
+    }
+
+    private void runEditAddress(Address address) {
+
+        boolean keepGoing = true;
+        String updateWith;
+        do {
+            //Display Current Address State
+            address = dao.getAddressById(address.getId());
+            view.printAddress(address);
+
+            //Update menu selection and handling
+            switch (view.printAndGetEditAddressMenuSelection()) {
+                case 1:
+                    updateWith = view.getField("First Name");
+                    dao.updateAddressFirstName(address.getId(), updateWith);
+                    break;
+                case 2:
+                    updateWith = view.getField("Last Name");
+                    dao.updateAddressLastName(address.getId(), updateWith);
+                    break;
+                case 3:
+                    updateWith = view.getField("Street Address");
+                    dao.updateAddressStreet(address.getId(), updateWith);
+                    break;
+                case 4:
+                    updateWith = view.getField("City");
+                    dao.updateAddressCity(address.getId(), updateWith);
+                    break;
+                case 5:
+                    updateWith = view.getField("State");
+                    dao.updateAddressState(address.getId(), updateWith);
+                    break;
+                case 6:
+                    updateWith = view.getField("Country");
+                    dao.updateAddressCountry(address.getId(), updateWith);
+                    break;
+                case 7:
+                    updateWith = view.getField("Post Code");
+                    dao.updateAddressPostCode(address.getId(), updateWith);
+                    break;
+                case 8:
+                    keepGoing = false;
+                    break;
+                default:
+                    break;
+            }
+        } while (keepGoing);
     }
 
 }
